Support ListDefinedTags and ChangeTagsListActive hooks
We define a tag if it's actually in use for a known consumer, and mark
it active if the consumer is proposed or approved. This avoids flooding
Special:Tags with tags for consumers that aren't taking any logged
actions.

Bug: T60312
Change-Id: I893bfdef051f03216787801d6ed4317917d74a11
(cherry picked from commit fa9daba73a83e7e4b54ba85960f31aa9d7c7e6ce)